new_bot/requirements.txt
python-telegram-bot==20.7
yookassa==3.0.0
sqlite3